Need myfft.m and myfilter.m function files
Both function files are required to .  Run the FINAL_Arduino.ino file in the IDE and ensure that the data is being recorded when through the serial monitor.  You can then run the ENGR131FINAL.mlapp file as long as the function files stated above are in the same folder, which you can check on the .m file in one of the tabs.